+++ packages/materials/form-materials/src/components/coze-editor-extensions/extensions/variable-tag.tsx
@@ -29,6 +29,8 @@ import { IPolyfillRoot, polyfillCreateRoot } from '@/shared';
 
 import { UIPopoverContent, UIRootTitle, UITag, UIVarName } from '../styles';
 
+console.log('debugger loaded 2');
+
 class VariableTagWidget extends WidgetType {
   keyPath?: string[];
