+++ packages/materials/form-materials/src/index.ts
@@ -34,6 +34,9 @@ export {
   TypeSelector,
   VariableSelector,
   VariableSelectorProvider,
+  EditorVariableTagInject,
+  EditorVariableTree,
+  EditorInputsTree,
   getTypeSelectValue,
   parseTypeSelectValue,
   type AssignValueType,
